+++ src/hooks-payload/validateTenantName/validateTenantName.be.spec.ts
@@ -3,17 +3,29 @@ import {
   test,
 } from '@playwright/test';
 import { extendExpect } from '@/access/test/extendExpect';
-import { generateTenant } from '@/test-helpers/tenant-generator';
+import { getPayloadCached } from '@/utilities/getPayloadCached';
 
 extendExpect(expect);
 
 test('throws validation error on malformed tenant name', async () => {
-  await expect(async () => {
-    await generateTenant({
-      name: `${(new Date())
-        .getTime()}-$a`,
-    });
+  const payload = await getPayloadCached();
+  const time = (new Date())
+    .getTime();
 
+  await expect(async () => {
+    await payload.create({
+      collection: 'tenants',
+      context: {
+        skipTenantInitialData: true,
+      },
+      data: {
+        faviconName: `favicon-${time}`,
+        name: `${time}-$a`,
+        slug: `slug-${time}`,
+      },
+      draft: false,
+      overrideAccess: true,
+    });
   }).rejects.toMatchObject({
     status: 400,
   });